+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## v0.1.4-dev
 
+- Hatter maps now perserve insertion order, unlike Rust but like
+  Ruby/JS/Lua.
 - More [builtins]: `pop()` for lists.
 - Finally added `set_index` operations for List, Map, and Object,
   both `a[1] = 2` and `map.thing = true` forms.